+++ lib/rack/robustness.rb
@@ -160,7 +160,7 @@ module Rack
 
     def handle_headers(ex, headers)
       handle_value(ex, headers).each_pair do |key,value|
-        @response[key] = handle_value(ex, value)
+        @response[key] ||= handle_value(ex, value)
       end
     end