+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.1.0 / 2013-04-11
 
 * Renamed `#on` as `#rescue` for better capturing semantics of `on` blocks (now an alias)
+* Added suppport for ensure clause(s), called after `rescue` blocks on every error
 
 # 1.0.0 / 2013-02-26